move ui_helper.ts
